Add mypy pre-commit hook (lax baseline on src/bolinas/)
mypy is configured in pyproject.toml under [tool.mypy] with
``ignore_missing_imports`` + ``strict_optional = False`` — a deliberately
lax baseline that respects the many stubless transitive deps (lightning,
torchmetrics, polars-bio, bioframe, cyvcf2, datasets, etc.) and the
codebase's implicit-None tolerance. Tighten incrementally in follow-ups.

The hook lives in `.pre-commit-config.yaml` as a local hook using
``uv run mypy`` (language: system) so it sees the actual project venv.
Triggers on any ``src/bolinas/**/*.py`` change and re-checks the whole
tree (single-file mypy misses cross-file issues).

Required 54 fixes across 12 files in `src/bolinas/pipelines/` to clear
the baseline. Categories — none changed runtime behavior:

- Polars schema annotations widened from ``dict[str, pl.DataType]`` to
  ``dict[str, type[pl.DataType] | pl.DataType]`` (the dict values are
  classes like ``pl.Utf8`` / ``pl.Int64``, not instances; polars
  accepts both at runtime).
- Lightning ``self.hparams.X`` switched to subscript access
  ``self.hparams["X"]`` (both work at runtime; Lightning types
  ``hparams`` as ``MutableMapping``).
- ``bin_feature``'s ``edges`` parameter widened from ``list[float]``
  to ``Sequence[float]`` to accept ``list[int]`` callers (covariance).
- ``assert not isinstance(optimizer, list)`` to narrow
  ``LightningOptimizer | list[LightningOptimizer]`` from
  ``self.optimizers()`` (holds when only one optimizer is configured).
- Targeted ``# type: ignore[...]`` comments for stub-vs-runtime
  mismatches: torchmetrics' ``compute()`` typed ``-> None``,
  ``nn.ModuleDict[key]`` returning bare ``Module``, ``torch.compile``
  returning a generic callable, polars ``Then``/``ChainedThen`` chain
  typing, Lightning's strict ``configure_optimizers`` return signature.

Library tier (``src/bolinas/data/``, ``model/``, ``tokenizer/``,
``levanter/``) was already mypy-clean before this commit — no changes
there. ``mypy>=1.18.0`` added to ``[dependency-groups].dev``.

879 pytest tests still pass.

+++ src/bolinas/pipelines/evals/plotting.py
@@ -133,7 +133,7 @@ def plot_models_comparison(
     output_path: str | Path,
     score_type: str | None = None,
     dataset_subset_score_map: dict[tuple[str, str], str] | None = None,
-    figsize: tuple[int, int] | None = None,
+    figsize: tuple[float, float] | None = None,
     models_filter: list[str] | None = None,
     dataset_subsets_filter: list[tuple[str, str]] | None = None,
     baseline_data: dict[tuple[str, str], dict[str, float]] | None = None,
